% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBANCOVA.RMST-package.R
\docType{package}
\name{RBANCOVA.RMST-package}
\alias{RBANCOVA.RMST}
\alias{RBANCOVA.RMST-package}
\title{RBANCOVA.RMST: Randomization-Based ANCOVA for RMST estimates with categorized time-to-event data}
\description{
Function for RB-ANCOVA with restricted mean survival time (RMST) using categorized time-to-event data.
}
\author{
\strong{Maintainer}: Taylor Krajewski \email{taylor.krajewski@duke.com}

}
\keyword{internal}
