% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBANCOVA_RMST.R
\name{rbancova_rmst}
\alias{rbancova_rmst}
\title{RB-ANCOVA for RMST}
\usage{
rbancova_rmst(
  data_treated,
  data_control,
  time_var,
  event_var,
  covariate_vars,
  timepoints,
  num_intervals
)
}
\arguments{
\item{data_treated}{A data frame subset to the treated units}

\item{data_control}{A data frame subset to the control units}

\item{event_var}{Name of the event indicator (0/1).}

\item{covariate_vars}{list of covariates to adjust for}

\item{timepoints}{Numeric vector of taus for RMST.}

\item{num_intervals}{Integer: number of intervals for discretization.}
}
\value{
A data frame with one row per difference estimate
}
\description{
RB-ANCOVA for RMST
}
